rotate to starting pose
This commit adds a behavior that rotates the robot into a starting pose prior to moving to the predocking position. This helps the robot dock successfully when it is on the left side of the docking station and facing forward. Previously, docking only worked when the robot was on the left side of the docking station and facing backward.

Being on the right side of the docking station is now detected and throws a NotImplementedError. When this is commented out, the robot can get close to docking successfully, but has not succeeded when tested.

A key part of all of these behaviors is that the robot needs to keep the docking station ArUco marker in view, but the robot's mast can occlude it. Thus, special consideration is required for directions of rotation and starting poses.