new 3D docking station coordinate system
This commit adds a 3D docking station coordinate system. Motivations include the following:

+ Simplify the creation of way points with respect to the docking station
+ Enable metric values to be used when creating way points with respect to the docking station
+ Provide an alternative to image-space approximations
+ Provide clearer mathematical justification for the approach
+ Eventually improve the performance of autonomous docking (not yet achieved)

The previous implementation used image-space features to approximate a coordinate system for the docking station. For example, the negative of the y-axis for the docking station's ArUco marker was used to generate a vector in the image that was approximately parallel to the midline of the docking station. This approximation improved as the robot approached the docking station enabling greater precision when it was most needed. It required a constant value to be tuned to compute how many pixels away from the docking station a way point should be.

Autonomous docking performance with the new 3D docking station coordinate system has degraded, but has not yet been tuned nor thoroughly tested. It worked on a couple of trials, but the way point was farther away from the docking station resulting in more error and hence more reliance on the mechanical guide that can direct the omniwheel.