refactoring & first success from right side of docking station
This commit refactors the code. It also adds a rotation behavior to point the robot toward the pre-docking waypoint prior to moving to the waypoint.

Although it needs more work, this code enabled the robot to successfully dock when starting from the right side of the docking station. This is the first time for success from the right side of the docking station.

This code has not been thoroughly tested.

One major change from refactoring the code is the use of a function for each of the behaviors.